ensure multiple markdown tables are rendered
This seeks to fix issue #61 by printing all output to the `io.Writer` passed to
`writer.TableWriter#Write`. This ensures multiple, distinct markdown tables are
printed, even when `tf-summarize` is invoked with both `-md` and `-out` flags.

Previously, `fmt.Println()` was used to separate the markdown tables,
but this is insufficient when `-out` is used with a non-STDOUT writer, as is the
case when writing output to a `-out`-specified file.