Fix issue #20 "unexpected end of JSON input" error (#64)
* correct input reading logic

This addresses https://github.com/dineshba/tf-summarize/issues/20 and
ensures input is read from the plan file -- and not via STDIN -- if a
plan file argument is provided.

This also seeks to improve some of the error messaging to be a bit more
clear.

* demo issue 20 fix via GH Actions

Ideally, tf-summarize would feature a suite of automated tests verifying
its functionality. In absenece of that, this demos the issue #20 fix via
GH Actions.

* use consistent TF version when invoking TF

- ensure the generation of the `example` directory data is done in a
  consistent, reproducible fashion
- ensure GH Actions uses the same version of TF expected by the `example` directory
- add plan and plan JSON files to source control, for testing purposes

* add automated test verifying STDIN vs file-provided input

This adds a basic automated test verifying the validity of the issue #20
fix.

* remove `demo` job from Build GH Actions workflow

Per code review request, @dineshba would prefer this be kept in a
separate file.