feat(table, HTML): include 'moved' resources (#97)
* feat(moved): GetAllResourceChanges includes 'moved'

This lays a foundation for including 'moved' in summarized plan out put,
ultimately to address issue #95

Signed-off-by: Mike Ball <mikedball@gmail.com>

* chore(html): do not overwrite templates dir in tests

Previously, https://github.com/dineshba/tf-summarize/pull/78 introduced
a change to overwrite the baked-in templates directory with a mock file
system during testing. As a result, tests did not exercise the real
`writer/templates/*.html` tf-summarize template logic/result.

Signed-off-by: Mike Ball <mikedball@gmail.com>

* fix(html): specify change type in CHANGE table column

Previously, the HTMLWriter tests' "CHANGE" column value was a resource
name, and not the actual type of change seen in the plan. This fixes
that.

Signed-off-by: Mike Ball <mikedball@gmail.com>

* feat(html): include "moved" resource changes

This includes resource address moves in the resource changes in HTML
table output.

Signed-off-by: Mike Ball <mikedball@gmail.com>

* feat(table): include "moved" resource details in MD table

This includes details pertaining resource moves in the markdown table output.

* feat(table): include "moved" resources in non-MD table

This includes moved resources amongst the changes in the non-markdown
table output.

Signed-off-by: Mike Ball <mikedball@gmail.com>

* feat(html): ensure 'moved' resources appear at bottom of change table

This improves the HTML table to show 'moved' resources at the bottom of
the table. This also seeks to establish a better control flow pattern
for other writers to emulate.

Signed-off-by: Mike Ball <mikedball@gmail.com>

* feat(table): pass/template moved resources separate from other changes

This seeks to enable better control/flow around moved-specific logic.

Signed-off-by: Mike Ball <mikedball@gmail.com>

* fix(GetAllResourceChanges): disinclude moves from resource changes

A move does not inherently require change actions; moves are handled
separately from resource changes.

Signed-off-by: Mike Ball <mikedball@gmail.com>

---------

Signed-off-by: Mike Ball <mikedball@gmail.com>